pyproject.toml
[project]
name = "v4vapp-backend-v2"
version = "0.5.12"
description = "Back End for the V4V.app bridge from Hive to Lightning"
authors = [{name = "Brian of London (home imac)", email = "brian@v4v.app"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "googleapis-common-protos>=1.62.0",
    "mypy-protobuf>=3.6.0,<4.0.0",
    "pydantic[email]>=2.7.1,<3.0.0",
    "colorlog>=6.8.2,<7.0.0",
    "backoff>=2.2.1,<3.0.0",
    "single-source>=0.3.0,<1.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "types-pyyaml>=6.0.12.20240917,<7.0.0",
    "httpx>=0.27.2,<1.0.0",
    "toml>=0.10.2,<1.0.0",
    "typer==0.16.0",
    "aiofiles>=24.1.0,<25.0.0",
    "watchdog>=6.0.0,<7.0.0",
    "asgiref>=3.9.1,<4.0.0",
    "colorama>=0.4.6,<1.0.0",
    "redis>=5.2.1,<6.0.0",
    "binance-connector>=3.12.0,<4.0.0",
    "python-telegram-bot>=22.3,<23.0",
    "hive-nectar>=0.0.11",
    "tenacity>=9.1.2",
    "packaging>=25.0",
    "bech32>=1.2.0",
    "pandas>=2.3.1",
    "tabulate>=0.9.0",
    "fastapi[standard]>=0.116.1",
    "uvicorn>=0.35.0",
    "types-tabulate>=0.9.0.20241207",
    "aiocache>=0.12.3",
    "pymongo==4.15.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4,<9.0.0",
    "pytest-asyncio>=0.23.7,<1.0.0",
    "types-pyyaml>=6.0.12.20240917,<7.0.0",
    "types-toml>=0.10.8.20240310,<1.0.0",
    "pytest-cov>=6.0.0,<7.0.0",
    "mypy>=1.15.0,<2.0.0",
    "black>=25.1.0,<26.0.0",
    "mongomock-motor>=0.0.35,<1.0.0",
    "pytest-mock>=3.14.0,<4.0.0",
    "types-aiofiles>=24.1.0.20241221,<25.0.0",
    "types-colorama>=0.4.15.20240311,<1.0.0",
    "dill>=0.3.9,<1.0.0",
    "hive-nectar>=0.0.3",
    "pre-commit>=4.2.0",
    "pandas-stubs>=2.2.3.250308",
    "matplotlib>=3.10.3",
    "jupyter",
    "ipykernel",
    "matplotlib>=3.10.3",
    "pandas-stubs>=2.2.3.250308",
    "pymongo>=4.13.0",
    "nbstripout>=0.8.1",
    "types-redis>=4.6.0.20241004",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/v4vapp_backend_v2"]

[tool.ruff]
line-length = 99
exclude = [
  "**/*_grpc.py",  # Excludes all files ending with _grpc.py
  "**/*_pb2.py",    # Excludes all files ending with _pb2.py
  "**/*.pyc",       # Excludes all compiled Python files
  "**/__pycache__/*" # Excludes all files in __pycache__ directories
]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
log_file = "pytest.log"
filterwarnings = [
    "ignore::DeprecationWarning:beem\\..*",  # Suppress warnings from the beem library
    "ignore::DeprecationWarning:pytz\\..*",  # Suppress warnings from the pytz library
    "ignore:.*Event loop is closed.*",  # Suppress event loop closed warnings
    "ignore:.*Task was destroyed but it is pending.*",  # Suppress task destroyed warnings
    "ignore:.*PytestUnraisableExceptionWarning.*"  # Suppress unraisable exception warnings
]

asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "module"
asyncio_default_test_loop_scope = "module"

[tool.mypy]
exclude = ".*typing.py|.*_pb2.*|.*.proto"  # Combine all exclusions into a single regex pattern

[tool.uv.sources]
hive-nectar = { git = "https://github.com/TheCrazyGM/hive-nectar.git" }
